preliminary calculator client working
A preliminary version of the calculator client (calc_client)
is working.  It can send expressions and recieve results.
Input can be taken on stdin either from the user or from a file.

This version does have a significant bug.
When the client finishes and exits, the server goes to
100% CPU usage.  And in some cases when a new client starts
and tries to send expressions it just gets stuck as if the
server was not responding.  More investigation is needed to
determine where the fault lies.