fixed bug, calc_client missed sending null
The calc_client wasn't properly sending a null terminated string.
The server (calc_server) assumes the string is properly terminated
so when it isn't it reads beyond the buffer, resulting in invalid
calculations.  Line 233 of the server shows that it has no length
restriction as a result of its use of sscanf().