unreliable sendto added to arq example
The unreliable sento (packetErrorSendTo) has been substituted
in for sendto in the client and server.  It appears to work
normally for a period and then it drops a packet.  No error
is reported by packetErrorSendTo.  And since a blocking recvfrom
is used it blocks forever waiting for a packet that will never
arrive.

    jeri@hudson:~/socket-examples/arq$ diff snw-client.c ../udp-echo/echo_client.c
    30,31d29
    < #include "packetErrorSendTo.h"
    <
    92c90
    < 			n = packetErrorSendTo(sockfd, buf+i, left, 0, res->ai_addr, res->ai_addrlen);
    ---
    > 			n = sendto(sockfd, buf+i, left, 0, res->ai_addr, res->ai_addrlen);
    jeri@hudson:~/socket-examples/arq$