preliminary ARQ functionality added to snw-client
A very simply ARQ scheme has been implemented in the
snw-client.  If it takes too long to get a reply it
times out and re-sends the packet.  As long as the timeout
is longer than the maximum latency for a normal RTT it
will work.  If it was too short there could be duplicate
packets.  An special ACK packet is not used but since the
server will simply echoes the data it receives, the same
effect is accomplished.