added packet sequence number
With the addition of a single bit sequence number
the send timeout can be reduced to zero and it will
still reliably transfer packets.  There will, however,
be lots of duplicates which make it slower.  This
can be configured using the TIMEOUT_MS and MAXSEQ variables
in snw-client.c

The DEBUG output can be enabled to show all sorts of operation
information and statistics.