arq: fix loop hang by resetting the sequence number
When the snw-server is changed to loop and allow multiple transfers it
would hang while attempting the second transfer.  This was caused by an
ARQ sequence number during starting which caused it to think the data
was invalid.  Resetting the sequence number for new connections fixes
the problem and allows for multiple transfers.