correct udp echo client use of getaddrinfo results
The echo client incorrectly used only the primary result (res)
from getaddrinfo instead of the actual result which was found
to be valid (p).  Correct this so it uses the valid result.

This change is similar to 2fca0d17c605 which corrected the
same problem but for the tcp echo client.